+++ core/src/main/kotlin/ca/allanwang/kau/utils/TransitionUtils.kt
@@ -1,7 +1,7 @@
 package ca.allanwang.kau.utils
 
-import android.support.transition.Transition
-import android.support.transition.TransitionSet
+import android.support.transition.Transition as SupportTransition
+import android.transition.Transition
 
 /**
  * Created by Allan Wang on 2017-06-24.
@@ -14,6 +14,18 @@ class TransitionEndListener(val onEnd: (transition: Transition) -> Unit) : Trans
     override fun onTransitionStart(transition: Transition) {}
 }
 
-@KauUtils fun TransitionSet.addEndListener(onEnd: (transition: Transition) -> Unit) {
+@KauUtils fun Transition.addEndListener(onEnd: (transition: Transition) -> Unit) {
     addListener(TransitionEndListener(onEnd))
+}
+
+class SupportTransitionEndListener(val onEnd: (transition: SupportTransition) -> Unit) : SupportTransition.TransitionListener {
+    override fun onTransitionEnd(transition: SupportTransition) = onEnd(transition)
+    override fun onTransitionResume(transition: SupportTransition) {}
+    override fun onTransitionPause(transition: SupportTransition) {}
+    override fun onTransitionCancel(transition: SupportTransition) {}
+    override fun onTransitionStart(transition: SupportTransition) {}
+}
+
+@KauUtils fun SupportTransition.addEndListener(onEnd: (transition: SupportTransition) -> Unit) {
+    addListener(SupportTransitionEndListener(onEnd))
 }